test: add Ray basic deduplicator test mapping
